Create Hokeemon
Turn-based game that allows player three options- Listen, Play, and Feed. If the player plays with his hokeemon creature, it becomes less bored, and if the player feeds his hokeemon creature, it becomes less hungry. The Listen function displays the hunger and boredom status. If your creature becomes too bored or too hungry, it will die and the game is over.